Refactor: 불 필요한 ArticlePreview 컴포넌트 레이어를 제거한다
