+++ src/components/home/ArticleSection/ArticleSection.tsx
@@ -1,6 +1,6 @@
-import { ArticlePreview } from '@/components';
-import { ArticleType } from '@/components/common/ArticlePreview/ArticlePreview';
+import { ArticlePreview, ArticlePreviewLg } from '@/components';
 import { useDetectViewport } from '@/hooks';
+import { ArticleType } from '@/components/common/ArticlePreview/ArticlePreview';
 import * as Styled from './ArticleSection.styled';
 
 interface ArticleSectionProps {
@@ -13,11 +13,9 @@ const ArticleSection = ({ allContentfulArticles }: ArticleSectionProps) => {
   const { viewportSize } = useDetectViewport();
   return (
     <Styled.ArticleSection>
-      {viewportSize !== 'mobile' && (
-        <ArticlePreview article={allContentfulArticles.nodes[0]} size="lg" />
-      )}
-      <ArticlePreview article={allContentfulArticles.nodes[0]} size="md" />
-      <ArticlePreview article={allContentfulArticles.nodes[1]} size="md" />
+      {viewportSize !== 'mobile' && <ArticlePreviewLg article={allContentfulArticles.nodes[0]} />}
+      <ArticlePreview article={allContentfulArticles.nodes[0]} />
+      <ArticlePreview article={allContentfulArticles.nodes[1]} />
     </Styled.ArticleSection>
   );
 };
