+++ src/components/home/index.ts
@@ -1,2 +1,3 @@
 export { default as Header } from './Header/Header';
+export { default as ArticlePreviewLg } from './ArticlePreviewLg/ArticlePreviewLg';
 export { default as ArticleSection } from './ArticleSection/ArticleSection';
